% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caps.R
\name{caps}
\alias{caps}
\title{Make Text All Caps or No Caps}
\usage{
caps(x, upper = TRUE)
}
\arguments{
\item{x}{text string}

\item{upper}{whether string should be upper or lower}
}
\description{
Make Text All Caps or No Caps.
}

