% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_kleen.R
\name{deep_kleen}
\alias{deep_kleen}
\title{Title Cleanse Text Of Common Errors}
\usage{
deep_kleen(xx)
}
\arguments{
\item{xx}{text string / vector}
}
\description{
Description Removes html tags, non-ascii, new-line characters, and encodes to UTF-8
}

