% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_char.R
\name{get_first_char}
\alias{get_first_char}
\title{Extracts the First Character}
\usage{
get_first_char(x)
}
\arguments{
\item{x}{text string / vector}
}
\description{
Extracts the first character of each word in text string.
}

