% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_char.R
\name{get_last_char}
\alias{get_last_char}
\title{Extract the Last Character}
\usage{
get_last_char(x)
}
\arguments{
\item{x}{text string}
}
\description{
Extract the last charcter from each word in a string.
}

