% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_seq.R
\name{get_num_seq}
\alias{get_num_seq}
\title{Extract a Sequence of Numbers from Text}
\usage{
get_num_seq(xx)
}
\arguments{
\item{xx}{text string / vector}
}
\description{
Extract a Sequence of Numbers from Text.
}

