% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kleentex-package.R
\docType{package}
\name{kleentex}
\alias{kleentex}
\alias{kleentex-package}
\alias{package-kleentex}
\title{What the Package Does (one line)}
\description{
What the package does (one or more paragraphs)
}

