% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_newline.R, R/rm_tabs.R
\name{rm_newline}
\alias{rm_newline}
\title{Remove Newline Characters}
\usage{
rm_newline(xx)

rm_newline(xx)
}
\arguments{
\item{xx}{text string / vector}

\item{xx}{text string / vector}
}
\description{
Remove Newline Characters.

Remove Tab-Delimiters.
}

