% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_nonalpha.R
\name{rm_nonalpha}
\alias{rm_nonalpha}
\title{Remove Non-Alpha Characters from Text}
\usage{
rm_nonalpha(xx)
}
\arguments{
\item{xx}{text string / vector}
}
\description{
Remove Non-Alpha Characters from Text.
}

