% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_stopwords.R
\name{rm_stopwords}
\alias{rm_stopwords}
\title{Remove Stopwords from Vector of Words}
\usage{
rm_stopwords(xx)
}
\arguments{
\item{xx}{vector of words to remove stopwords from}
}
\description{
Removes Stopwords from Vector of Words.
}

