% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_words.R
\name{stem_words}
\alias{stem_words}
\title{Stem a Vector of Words}
\usage{
stem_words(xx, language = "porter")
}
\arguments{
\item{xx}{text vector}

\item{language}{which algorithm to use}
}
\description{
Stem a Vector of Texts.
}

